% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.r
\name{remove_high_missing_col}
\alias{remove_high_missing_col}
\title{Remove Columns with a High Proportion of Missing Values}
\usage{
remove_high_missing_col(x, prop = 1, n = 1000)
}
\arguments{
\item{x}{a dataframe}

\item{prop}{a number between 0 and 1, if a column has more than this proportion of missing values it is removed. (default = 1)}

\item{n}{the number of rows to use to sample from when calculating the proportion of missing vaalue. useful for large datasets. (default = 1000)}
}
\value{
a \code{dataframe} with missing rows removed
}
\description{
Use \code{remove_high_missing_col} to remove columns from a dataframe that have an excessive proportion of missing values.
}
