% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.r
\name{remove_high_missing_row}
\alias{remove_high_missing_row}
\title{Remove Rows with a High Proportion of Missing Values}
\usage{
remove_high_missing_row(x, prop = 1, n = 1000)
}
\arguments{
\item{x}{a dataframe}

\item{prop}{a number between 0 and 1. If a row has more or equal to this proportion of missing values it is removed. (default = 1)}

\item{n}{the number of columns to use to sample from when calculating the proportion of missing vaalue. useful for large datasets. (default = 1000)}
}
\value{
a \code{dataframe} with missing rowss removed
}
\description{
Use \code{remove_high_missing_row} to remove rows from a datarame that have an excessive proportion of missing values.
}
