% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-method.r
\name{impute}
\alias{impute}
\alias{impute.numeric}
\alias{impute.character}
\alias{impute.factor}
\alias{impute.logical}
\alias{impute.data.frame}
\title{Simple Imputation}
\usage{
impute(x, ...)

\method{impute}{numeric}(x, .fun = median, ...)

\method{impute}{character}(x)

\method{impute}{factor}(x)

\method{impute}{logical}(x)

\method{impute}{data.frame}(x, ...)
}
\arguments{
\item{x}{Numeric, Factor, Logical or Charater vector, or a Data Frame}

\item{...}{arugments to be passed to othermethods}

\item{.fun}{a fuction (default median) to use for imputation}
}
\description{
Generic simple Imputation method
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Median Imputation for a numeric Vector

\item \code{character}: Mode Imputation for a character vector

\item \code{factor}: Mode Imputation for a factor vector

\item \code{logical}: Mode Imputation for a logical vector

\item \code{data.frame}: ddata frame imputation
}}

